Moving metrics to their own file.
And other cleanups.
